requirements.txt
PyOpenGL==3.1.0
glfw==1.4.0
pyrr==0.8.3
Pillow==4.1.1
pyrocket==0.2.5
PyWavefront==0.3.1

# Optional
#pygame==1.9.3
